+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.23.0
 
+* More progress indicators, and `-q,--quiet` options to turn them off
+* `doc` command implies `sync` command
 * Add `sync` command
 * `-f,--force` option for `undiff`, `repo`, and `unrepo` commands
 * Proper command line option parsing, better help system